backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
setuptools==80.3.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
wheel==0.45.1
